% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertify.R
\name{alertify_alert}
\alias{alertify_alert}
\title{Alertify Alerts}
\usage{
alertify_alert(
  title = "Alert Title",
  message = "Alert Message",
  type = "success",
  btn_label = "OK",
  transition = "pulse",
  transition_off = FALSE,
  closable = TRUE,
  auto_reset = FALSE,
  frameless = FALSE,
  maximizable = FALSE,
  modal = FALSE,
  movable = FALSE,
  move_bounded = TRUE,
  overflow = FALSE,
  padding = TRUE,
  pinnable = FALSE,
  resizeable = FALSE,
  start_maximized = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{title}{Dialog title.}

\item{message}{Dialog contents.}

\item{type}{Dialog type. Defaults to `"success"`. Valid values are: 
\itemize{
\item `"success"`
\item `"error"`
\item `"warning"`
\item `"message"`
}}

\item{btn_label}{The `OK` button label.}

\item{transition}{Transition effect to be used when showing/hiding the dialog. Defaults to `"pulse"`. Valid values are: 
\itemize{
\item `"pulse"`
\item `"slide"`
\item `"zoom"`
\item `"fade"`
\item `"flipx"`
\item `"flipy"`
}}

\item{transition_off}{Logical; if `TRUE`, transition effect is disabled. Defaults to `FALSE`.}

\item{closable}{Logical; if `TRUE` (the default), a `Close` button is displayed in the header of the dialog.}

\item{auto_reset}{Logical; if `TRUE` (the default), the dialog will reset size/position on window resize.}

\item{frameless}{Logical; if `TRUE`, hides both header and footer of the dialog. Defaults to `FALSE`.}

\item{maximizable}{Logical; if `TRUE` (the default), the `Maximize` button is displayed in the header of the dialog.}

\item{modal}{Logical; if `TRUE` (the default), a screen dimmer will be used and access to the page contents will be prevented.}

\item{movable}{Logical; if `TRUE` (the default), the dialog is movable.}

\item{move_bounded}{Logical; if `TRUE`, the dialog is not allowed to go off-screen. Defaults to `FALSE`.}

\item{overflow}{Logical; if `TRUE` (the default), the content overflow is managed by the dialog}

\item{padding}{Logical; if `TRUE` (the default), the content padding is managed by the dialog.}

\item{pinnable}{Logical; if `TRUE` (the default), the `Pin`` button is displayed in the header of the dialog.}

\item{resizeable}{Logical; if `TRUE`, the dialog is resizable. Defaults to `FALSE`.}

\item{start_maximized}{Logical; if `TRUE`, the dialog will start in a maximized state. Defaults to `FALSE`.}
}
\description{
Pretty browser dialogs.
}
