% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertify.R
\name{alertify_notify}
\alias{alertify_notify}
\title{Alertify Notifications}
\usage{
alertify_notify(
  message = "Alert Message",
  type = "success",
  delay = 5,
  position = "bottom-right",
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{message}{Notification message contents.}

\item{type}{Type of notification message. Defaults to `"success"`. Valid values are:
\itemize{
\item `"notify"`
\item `"mesage"`
\item `"success"`
\item `"warning"`
\item `"error"`
}}

\item{delay}{The time (in seconds) to wait before the notification is auto-dismissed. `0` will keep notification open till clicked.}

\item{position}{Position of the notification. Defaults to `"bottom-right"`. Valid values are:
\itemize{
\item `"bottom-right"`
\item `"bottom-left"`
\item `"bottom-center"`
\item `"top-right"`
\item `"top-left"`
\item `"top-center"`
}}
}
\description{
Pretty browser notifications.
}
