% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootbox.R
\name{bootBox}
\alias{bootBox}
\title{BootBox}
\usage{
bootBox(
  title = "Your title",
  message = "Your message here...",
  size = "small",
  close_on_escape = TRUE,
  show = TRUE,
  backdrop = NULL,
  close_button = TRUE,
  animate = TRUE,
  class = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{title}{Adds a header to the dialog.}

\item{message}{Text displayed in the dialog.}

\item{size}{Adds the relevant Bootstrap modal size class to the dialog wrapper. Valid values are:
\itemize{
\item `"small"`
\item `"large"`
\item `"extra-large"`
}}

\item{close_on_escape}{Logical; if `TRUE` (the default), allows the user to dismiss the dialog by hitting ESC key.}

\item{show}{Logical; if `TRUE` (the default), the dialog is shown immediately.}

\item{backdrop}{Logical; if `TRUE`, the backdrop is displayed and clicking on it dismisses the dialog. Defaults to `NULL`. Valid values are: 
\itemize{
\item `NULL`: The backdrop is displayed, but clicking on it has no effect.
\item `TRUE`: The backdrop is displayed, and clicking on it dismisses the dialog.
\item `FALSE`: The backdrop is not displayed.
}}

\item{close_button}{Logical; if `TRUE` (the default), a close button is displayed.}

\item{animate}{Logical; if `TRUE` (the default), animates the dialog in and out.}

\item{class}{Custom CSS class (using Animate.css).}
}
\description{
Bootstrap modals made easy.
}
