% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buzz.R
\docType{package}
\name{buzz}
\alias{buzz}
\title{\code{buzz} package}
\description{
Simple notifications for Shiny.
}
\details{
See the README on 
\href{https://github.com/rsquaredacademy/buzz}{GitHub}
}
\keyword{internal}
