% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtip.R
\name{microTip}
\alias{microTip}
\title{MicroTip}
\usage{
microTip(
  id = NULL,
  tip = "Hey! tooltip!",
  position = "top",
  size = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{The id of the element to attach the tooltip.}

\item{tip}{Content of the tooltip.}

\item{position}{Where the tooltip should appear relative to the target element. Defaults to `"top"`. Other valid     #' values are: 
\itemize{
\item `"bottom"`
\item `"left"`
\item `"right"`
\item `"top-left"`
\item `"top-right"`
\item `"bottom-left"`
\item `"bottom-right"`
}}

\item{size}{Size of the tooltip. Defaults to `"fit"` as the tooltip will takeup only the size it requires to show the #' text. Other valid values are:
\itemize{
\item `"small"`
\item `"medium"`
\item `"large"`
}}
}
\description{
Minimal CSS only tooltip.
}
