% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{previewStandby}
\alias{previewStandby}
\title{Preview Standby}
\usage{
previewStandby(type = "toast")
}
\arguments{
\item{type}{Type of alert/notification. Valid values are:
\itemize{
\item alertify
\item bootbox
\item notice
\item notify
\item tingle
\item toast
}}
}
\description{
Preview different types of alerts/notifications.
}
