% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tingle.R
\name{tingle}
\alias{tingle}
\title{Tingle}
\usage{
tingle(
  content = "Hello",
  close_button = FALSE,
  button_label = "Close",
  button_type = "default",
  button_position = "right",
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{content}{Content of the modal.}

\item{close_button}{Logical; if `TRUE`, displays a button to close the modal. Defaults to `FALSE`.}

\item{button_label}{Label of `close_button`.}

\item{button_type}{Type of button. Defaults to `"default"`. Other valid values are: 
\itemize{
\item `"primary"`
\item `"danger"`
}}

\item{button_position}{Position of the button inside the modal. Defaults to `"right"`. Valid values are: 
\itemize{
\item `"right"`
\item `"left"`
}}
}
\description{
Minimalist and easy to use modals.
}
