% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R, R/ns.R
\name{useNS}
\alias{useNS}
\alias{notice}
\title{Notification Styles}
\usage{
useNS()

notice(
  message = "Hello",
  type = "notice",
  layout = "growl",
  effect = "jelly",
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{message}{Notification message.}

\item{type}{Notification type. Default is `"notice"`. Other valid values are: 
\itemize{
\item `"success"`
\item `"warning"`
\item `"error"`
}}

\item{layout}{Notification layout. Default is `"growl"`. Other valid values are: 
\itemize{
\item `"attached"`
\item `"bar"`
}}

\item{effect}{Notification effect type. Valid values include:
\itemize{
\item For `"growl" layout`
\itemize{
\item `"scale"`
\item `"jelly"`
\item `"slide"`
\item `"genie"` 
} 
\item For `"attached" layout`
\itemize{
\item `"flip"`  
\item `"bouncyflip"`
}
\item For `"bar" layout`
\itemize{
\item `"slidetop"` 
\item `"exploader"` 
}
}}

\item{session}{Shiny session object.}
}
\description{
Simple website notifications with effects
}
\section{Functions}{

\itemize{
\item \code{useNS}: Dependencies to include in your UI.
\item \code{notice}: Display notifications.
}
}

